Update README.md
The page is a bit long. I think a preview in the form of a TOC would be nice. Also, the first thing we tell contributors is to use TOCs in their articles, so a bit ironic that this page itself doesn't have one :). Also, cleaned up some more text, and added free screencast tools for Linux and OSX