Update aws-cli.md
I couldn't find specific instructions on creating prefixes in the AWS documentation or this wiki, so I played around with it for a bit and figured this out. I thought it might be useful info to add to the wiki! My apologies if my wording or terminology is not exactly correct, I'm fairly new to S3 and still learning.

I work at Fred Hutch in the Meshinchi lab, by the way! My GitHub account uses a personal email address, though.