remove absolute links to the wiki
We are discussing having a CI/CD system to build non-main branches and deploy
them to some place where we can view them. For that and other reasons,
we want to remove absolute links so if we are looking at a branch hosted
somewhere, we don't want to accidentally click into the actual wiki
without realizing it.

Maybe in future we can add some functionality to CI/CD which checks for absolute links and either fixes it or throws an error during the build.