changed filename to be consistent with other files in the dir
Note apparent bug:
With a filename like

_scicompannounce/2022-08-15-rhino-gizmo-updates.md

if the date at the beginning of the filename is in the future, the file will not
be built/rendered, even though the index page will excerpt and link it.
Should figure out how to change this behavior. It's good to have dates in the future
in the URL for announcements of future maintenance.